requirements.txt
# Core dependencies for Jetson Nano Air Quality AI
numpy==1.21.6
pandas==1.5.3
scikit-learn==1.0.2
tensorflow==2.8.4
matplotlib==3.5.3
seaborn==0.11.2
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
flask==2.3.3
pyserial==3.5
schedule==1.2.0
python-dotenv==1.0.0
psutil==5.9.5

# For Jetson Nano GPIO (if needed)
Jetson.GPIO==2.0.21

# Additional utilities
tqdm==4.66.1
joblib==1.3.2
